test: replace ManyToManyAutoRelationIT.kt with ManyToManyIT
